moved thumbnails cache to thumbnails model
